Convert jellyfin-api to Kotlin Multiplatform
