+++ jellyfin-api/src/commonTest/kotlin/org/jellyfin/sdk/api/client/AuthorizationHeaderBuilderTests.kt
