+++ jellyfin-api/src/jvmMain/kotlin/org/jellyfin/sdk/api/client/extensions/MediaInfoApiExtensions.kt
