+++ jellyfin-api/src/jvmMain/kotlin/org/jellyfin/sdk/api/sockets/SocketSubscription.kt
