+++ jellyfin-api/src/jvmMain/kotlin/org/jellyfin/sdk/api/sockets/WebSocketApi.kt
